Game_analytics/requirements.txt
streamlit
pandas
mysql-connector-python
python-dotenv
requests
